requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
python-dotenv==0.19.0